Mongoose 적용 후 main과 병합

+++ controller/post.mjs
@@ -22,8 +22,8 @@ export async function getPost(req, res, next) {
 
 // 새로운 포스트 작성
 export async function createPost(req, res, next) {
-  const { userid, name, text } = req.body;
-  const posts = await postRepository.create(userid, name, text);
+  const { text } = req.body;
+  const posts = await postRepository.create(text, req.id);
   res.status(200).json(posts);
 }
 
@@ -31,17 +31,28 @@ export async function createPost(req, res, next) {
 export async function updatePost(req, res, next) {
   const id = req.params.id;
   const text = req.body.text;
-  const update = await postRepository.update(id, text);
-  if (update) {
-    res.status(201).json(update);
-  } else {
-    res.status(404).json({ message: `${id}의 포스트가 없습니다.` });
+  const post = await postRepository.getById(id);
+  if (!post) {
+    return res.status(404).json({ message: `${id}의 포스트가 없습니다.` });
   }
+  if (post.userId !== req.id) {
+    return res.sendStatus(403);
+  }
+  const updated = await postRepository.update(id, text);
+  res.status(200).json(updated);
 }
 
 // 포스트 삭제
 export async function removePost(req, res, next) {
   const id = req.params.id;
+  const post = await postRepository.getById(id);
+  if (!post) {
+    return res.status(404).json({ message: `${id}의 포스트가 없습니다.` });
+  }
+  if (post.userId !== req.id) {
+    return res.sendStatus(403);
+  }
+
   const remove = await postRepository.remove(id);
   res.status(204).json(`${id}번 글이 삭제되었습니다.`);
 }
